Major API change: reconcile.Result -> components.Result.
To better handle status management (and future growth), components now return a new struct called components.Result. It has the same fields as a reconcile.Result with added stuff specific to components. This includes a status generator function. This is returned as a function to allow it to be replayed in the event of a write collision. This also removes the need for the Statuser interface and code generator.